Move lambda tests to own file
